Fixed some bugs in should_set_tablename.
